refactor: 연동 페이지 각 컴포넌트 분리
